ADD: added cycloid tests to tests/_cycloid.py
